--- Assets/Audios.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 593d3f1efee0447e0ac093a53a36e5b3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
