+++ Assets/Prefabs/Grenade.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2745096beb24d4cc9ba325dd053e1ec2
+guid: 5fdec23ec9694486db73dbf5bc7ad389
 PrefabImporter:
   externalObjects: {}
   userData:
