+++ Assets/Scripts/BaseController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bb6b77ed8f76f4b4ab270c84961e5fa5
+guid: dc567e9944c03432aa0548a864195555
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
