+++ Assets/Scripts/ShotgunController.cs.meta
@@ -1,3 +1,3 @@
 ﻿fileFormatVersion: 2
-guid: 54e2dbfafd6e24898b1c1b596a8be168
+guid: 3f76941d510342028194e3f04d99d8db
 timeCreated: 1604276676
