--- Assets/Submissions/MissileController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 83912665e3fd4de199b09082b841a28a
-timeCreated: 1605986441
