--- Assets/Submissions/MissileLauncherController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 058945b4be8b435d90fe86b3ba496219
-timeCreated: 1605986285
