--- Assets/Submissions/MissileLauncherTest.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f88ea128120e44624b5c3c9198f528f8
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
